Let Composer figure out what version to install
This is more forward compatible (no need to change this line each time
you release a new major version), and also less error-prone, because
Composer knows this should go under require. Finally Composer will use
the caret operator instead of the tilde operator, or whatever the
correct operator is at the time it is executed.